requirements.txt
asgiref==3.2.3
Django==3.0
django-crispy-forms==1.8.1
numpy==1.17.4
Pillow==6.2.1
pytz==2019.3
scipy==1.3.2
sqlparse==0.3.0
virtualenv==16.7.8
certifi==2018.10.15
chardet==3.0.4
idna==2.7
requests==2.19.1
urllib3==1.23